Refactoring ahead of react-redux
